retire gpumonitor module v1, no longer needed (superseded)
