+++ AutoDarkModeSvc/Modules/WardenModule.cs
@@ -35,7 +35,7 @@ namespace AutoDarkModeSvc.Modules
             AutoManageModule(typeof(GeopositionUpdateModule), true, config.Location.Enabled);
             AutoManageModule(typeof(TimeSwitchModule), true, config.AutoThemeSwitchingEnabled && !State.PostponeSwitch);
             AutoManageModule(typeof(ThemeUpdateModule), true, config.WindowsThemeMode.Enabled);
-            AutoManageModule(typeof(GPUMonitorModuleV2), true, config.GPUMonitoring.Enabled);
+            AutoManageModule(typeof(GPUMonitorModule), true, config.GPUMonitoring.Enabled);
             AutoManageModule(typeof(UpdaterModule), true, config.Updater.Enabled);
         }
